inherit controlled gates from Controlled
