+++ pennylane/ops/math/control.py
@@ -14,12 +14,13 @@
 # pylint: disable=too-many-arguments
 
 from multimethod import multimethod
+import warnings
 
 import pennylane as qml
 from pennylane.operation import Operator, AnyWires
 from pennylane.wires import Wires
 from pennylane import numpy as np
-from ..qubit import PauliY, SWAP, PhaseShift, RX, RY, RZ, Rot
+from ..qubit import PauliX, PauliY, SWAP, PhaseShift, RX, RY, RZ, Rot
 
 class NoDecompositionShortcut(Exception):
     """decomposition undefined."""
@@ -104,6 +105,74 @@ def c1_decomp(op: Rot):
             qml.RZ(omega, wires=wires[1]),
         ]
 
+def multicx_decomposition(control_wires, target_wire, work_wires):
+    num_work_wires_needed = len(control_wires) - 2
+    if len(work_wires) < num_work_wires_needed:
+        tot_wires = len(control_wires) + 2
+        partition = int(qml.numpy.ceil(tot_wires / 2))
+
+        first_part = control_wires[:partition]
+        second_part = control_wires[partition:]
+
+        gates = [
+            Controlled(qml.PauliX(work_wires[0]),
+                control_wires=first_part,
+                work_wires=second_part + target_wire,
+            ),
+            Controlled(qml.PauliX(target_wire),
+                control_wires=second_part + work_wires[0:1],
+                work_wires=first_part,
+            ),
+            Controlled(qml.PauliX(work_wires[0]),
+                control_wires=first_part,
+                work_wires=second_part + target_wire,
+            ),
+            Controlled(qml.PauliX(target_wire),
+                control_wires=second_part + work_wires[0:1],
+                work_wires=first_part,
+            ),
+        ]
+        return gates
+    # decomposition with many wires
+    work_wires = work_wires[:num_work_wires_needed]
+
+    work_wires_reversed = list(reversed(work_wires))
+    control_wires_reversed = list(reversed(control_wires))
+
+    gates = []
+
+    for i in range(len(work_wires)):
+        ctrl1 = control_wires_reversed[i]
+        ctrl2 = work_wires_reversed[i]
+        t = target_wire if i == 0 else work_wires_reversed[i - 1]
+        gates.append(qml.Toffoli(wires=[ctrl1, ctrl2, t]))
+
+    gates.append(qml.Toffoli(wires=[*control_wires[:2], work_wires[0]]))
+
+    for i in reversed(range(len(work_wires))):
+        ctrl1 = control_wires_reversed[i]
+        ctrl2 = work_wires_reversed[i]
+        t = target_wire if i == 0 else work_wires_reversed[i - 1]
+        gates.append(qml.Toffoli(wires=[ctrl1, ctrl2, t]))
+
+    for i in range(len(work_wires) - 1):
+        ctrl1 = control_wires_reversed[i + 1]
+        ctrl2 = work_wires_reversed[i + 1]
+        t = work_wires_reversed[i]
+        gates.append(qml.Toffoli(wires=[ctrl1, ctrl2, t]))
+
+    gates.append(qml.Toffoli(wires=[*control_wires[:2], work_wires[0]]))
+
+    for i in reversed(range(len(work_wires) - 1)):
+        ctrl1 = control_wires_reversed[i + 1]
+        ctrl2 = work_wires_reversed[i + 1]
+        t = work_wires_reversed[i]
+        gates.append(qml.Toffoli(wires=[ctrl1, ctrl2, t]))
+
+    return gates
+    
+
+
 class Controlled(Operator):
     r"""Wrapper denoting a controlled operation.
 
@@ -145,7 +214,10 @@ class Controlled(Operator):
     num_wires = AnyWires
 
     def __init__(self, base, control_wires, control_values=None, work_wires=None, do_queue=True, id=None):
-        
+        if isinstance(control_values, str):
+            warnings.warn("Specifying control values as a string is deprecated. Please use Sequence[Bool]", UserWarning)
+            control_values = [(x=="1") for x in control_values]
+
         self.base = base
         self._control_wires = Wires(control_wires)
         self._work_wires = Wires([]) if work_wires is None else Wires(work_wires)
@@ -224,7 +296,7 @@ class Controlled(Operator):
 
     @staticmethod
     def compute_decomposition(*args, wires=None, base=None, control_wires=None, control_values=None,
-        work_wires=None, **kwargs):
+        work_wires=None):
 
         flips = [qml.PauliX(w) for w, val in zip(control_wires, control_values) if not val]
         if len(control_wires) == 1:
@@ -232,10 +304,13 @@ class Controlled(Operator):
                 return flips+c1_decomp(base, control_wires)+flips
             except NoDecompositionShortcut:
                 pass
+        if len(control_wires) > 2 and isinstance(base, PauliX):
+            return flips + multicx_decomposition(control_wires, base.wires, work_wires) + flips
         return qml.operation.DecompositionUndefinedError
 
     @staticmethod
     def compute_eigvals(*args, **kwargs):
+        # what for this?
         raise qml.operation.EigvalsUndefinedError
 
     def generator(self):
